move profiles into datatype folder
