+++ post-processing/datatypes/function_profile.py
@@ -1,0 +1,78 @@
+# Copyright 2022 Fuzz Introspector Authors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""Reads the data output from the fuzz introspector LLVM plugin."""
+
+import os
+import copy
+import json
+import logging
+
+from typing import (
+    Any,
+    Dict,
+    List,
+    Optional,
+    Set,
+    Tuple,
+)
+
+import fuzz_cfg_load
+import fuzz_constants
+import fuzz_cov_load
+import fuzz_utils
+
+from exceptions import DataLoaderError
+
+logger = logging.getLogger(name=__name__)
+logger.setLevel(logging.INFO)
+
+
+class FunctionProfile:
+    """
+    Class for storing information about a given Function
+    """
+    def __init__(self, elem: Dict[Any, Any]) -> None:
+        self.function_name = elem['functionName']
+        self.function_source_file = elem['functionSourceFile']
+        self.linkage_type = elem['linkageType']
+        self.function_linenumber = elem['functionLinenumber']
+        self.return_type = elem['returnType']
+        self.arg_count = elem['argCount']
+        self.arg_types = elem['argTypes']
+        self.arg_names = elem['argNames']
+        self.bb_count = elem['BBCount']
+        self.i_count = elem['ICount']
+        self.edge_count = elem['EdgeCount']
+        self.cyclomatic_complexity = elem['CyclomaticComplexity']
+        self.functions_reached = elem['functionsReached']
+        self.function_uses = elem['functionUses']
+        self.function_depth = elem['functionDepth']
+        self.constants_touched = elem['constantsTouched']
+        self.branch_profiles = self.load_func_branch_profiles(elem['BranchProfiles'])
+
+        # These are set later.
+        self.hitcount: int = 0
+        self.reached_by_fuzzers: List[str] = []
+        self.incoming_references: List[str] = []
+        self.new_unreached_complexity: int = 0
+        self.total_cyclomatic_complexity: int = 0
+
+    def load_func_branch_profiles(self, yaml_branch_profiles: Any) -> Dict[str, BranchProfile]:
+        bp_loaded = {}
+        for entry in yaml_branch_profiles:
+            new_branch = BranchProfile()
+            new_branch.assign_from_yaml_elem(entry)
+            bp_loaded[new_branch.branch_pos] = new_branch
+
+        return bp_loaded
