+++ post-processing/fuzz_data_loader.py
@@ -38,641 +38,6 @@ logger = logging.getLogger(name=__name__)
 logger.setLevel(logging.INFO)
 
 
-class InputBug:
-    """
-    Holds data about a given bug found by fuzzers.
-    """
-    def __init__(
-        self,
-        source_file: str,
-        source_line: str,
-        function_name: str,
-        fuzzer_name: str,
-        description: str,
-        bug_type: str
-    ) -> None:
-        self.source_file = source_file
-        self.source_line = source_line
-        self.function_name = function_name
-        self.fuzzer_name = fuzzer_name
-        self.description = description
-        self.bug_type = bug_type
-
-
-class BranchProfile:
-    """
-    Class for storing information about conditional branches collected by LLVM pass.
-    """
-    def __init__(self) -> None:
-        self.branch_pos = str()
-        self.branch_true_side_pos = str()
-        self.branch_false_side_pos = str()
-        self.branch_true_side_reachable_complexity = -1
-        self.branch_false_side_reachable_complexity = -1
-        self.branch_true_side_not_covered_complexity = -1
-        self.branch_false_side_not_covered_complexity = -1
-        self.branch_true_side_hitcount = -1
-        self.branch_false_side_hitcount = -1
-        self.branch_true_side_funcs: List[str] = []
-        self.branch_false_side_funcs: List[str] = []
-
-    def assign_from_yaml_elem(self, elem: Dict[Any, Any]) -> None:
-        # This skips the path, as it may cause incosistancy vs coverage file names path
-        self.branch_pos = elem['Branch String'].split('/')[-1]
-        self.branch_true_side_pos = elem['Branch Sides']['TrueSide']
-        self.branch_false_side_pos = elem['Branch Sides']['FalseSide']
-        self.branch_true_side_funcs = elem['Branch Sides']['TrueSideFuncs']
-        self.branch_false_side_complexity = elem['Branch Sides']['FalseSideFuncs']
-
-    def assign_from_coverage(self, true_count: str, false_count: str) -> None:
-        self.branch_true_side_hitcount = int(true_count)
-        self.branch_false_side_hitcount = int(false_count)
-
-    def dump(self) -> None:
-        """
-        For debugging purposes, may be removed later.
-        """
-        print(self.branch_pos, self.branch_true_side_pos, self.branch_false_side_pos,
-              self.branch_true_side_reachable_complexity,
-              self.branch_false_side_reachable_complexity,
-              self.branch_true_side_not_covered_complexity,
-              self.branch_false_side_not_covered_complexity,
-              self.branch_true_side_hitcount, self.branch_true_side_hitcount)
-
-
-class FunctionProfile:
-    """
-    Class for storing information about a given Function
-    """
-    def __init__(self, elem: Dict[Any, Any]) -> None:
-        self.function_name = elem['functionName']
-        self.function_source_file = elem['functionSourceFile']
-        self.linkage_type = elem['linkageType']
-        self.function_linenumber = elem['functionLinenumber']
-        self.return_type = elem['returnType']
-        self.arg_count = elem['argCount']
-        self.arg_types = elem['argTypes']
-        self.arg_names = elem['argNames']
-        self.bb_count = elem['BBCount']
-        self.i_count = elem['ICount']
-        self.edge_count = elem['EdgeCount']
-        self.cyclomatic_complexity = elem['CyclomaticComplexity']
-        self.functions_reached = elem['functionsReached']
-        self.function_uses = elem['functionUses']
-        self.function_depth = elem['functionDepth']
-        self.constants_touched = elem['constantsTouched']
-        self.branch_profiles = self.load_func_branch_profiles(elem['BranchProfiles'])
-
-        # These are set later.
-        self.hitcount: int = 0
-        self.reached_by_fuzzers: List[str] = []
-        self.incoming_references: List[str] = []
-        self.new_unreached_complexity: int = 0
-        self.total_cyclomatic_complexity: int = 0
-
-    def load_func_branch_profiles(self, yaml_branch_profiles: Any) -> Dict[str, BranchProfile]:
-        bp_loaded = {}
-        for entry in yaml_branch_profiles:
-            new_branch = BranchProfile()
-            new_branch.assign_from_yaml_elem(entry)
-            bp_loaded[new_branch.branch_pos] = new_branch
-
-        return bp_loaded
-
-
-class FuzzerProfile:
-    """
-    Class for storing information about a given Fuzzer.
-
-    This class essentially holds data corresponding to the output of run of the LLVM
-    plugin. That means, the output from the plugin for a single fuzzer.
-    """
-    def __init__(
-        self,
-        filename: str,
-        data_dict_yaml: Dict[Any, Any],
-        target_lang: str = "c-cpp"
-    ) -> None:
-        self.introspector_data_file = filename
-        self.function_call_depths = fuzz_cfg_load.data_file_read_calltree(filename)
-        self.fuzzer_source_file: str = data_dict_yaml['Fuzzer filename']
-        self.binary_executable: str = ""
-        self.coverage: Optional[fuzz_cov_load.CoverageProfile] = None
-        self.file_targets: Dict[str, Set[str]] = dict()
-        self.target_lang = target_lang
-
-        # Create a list of all the functions.
-        self.all_class_functions = dict()
-        for elem in data_dict_yaml['All functions']['Elements']:
-            # Check if there is normalisation issue and log if so
-            if "." in elem['functionName']:
-                split_name = elem['functionName'].split(".")
-                if split_name[-1].isnumeric():
-                    logger.info(
-                        f"We may have a non-normalised function name: {elem['functionName']}"
-                    )
-
-            func_profile = FunctionProfile(elem)
-            logger.debug(f"Adding {func_profile.function_name}")
-            self.all_class_functions[func_profile.function_name] = func_profile
-
-    def resolve_coverage_link(
-        self,
-        cov_url: str,
-        source_file: str,
-        lineno: int,
-        function_name: str
-    ) -> str:
-        """Resolves a link to a coverage report."""
-
-        # For C/CPP
-        if self.target_lang == "c-cpp":
-            return cov_url + source_file + ".html#L" + str(lineno)
-
-        # For Python
-        logger.debug(f"Can't get link for {cov_url} -- {source_file} -- {lineno}")
-
-        # Temporarily for debugging purposes. TODO: David remove this later
-        # Find the html_status.json file. This is a file generated by the Python
-        # coverate utility and contains mappings from source to html file. We
-        # need this mapping in order to create links from the data extracted
-        # during AST analysis, as there we only have the source code.
-        html_summaries = fuzz_utils.get_all_files_in_tree_with_regex(".", ".*html_status.json$")
-        logger.info(str(html_summaries))
-        if len(html_summaries) > 0:
-            html_idx = html_summaries[0]
-            with open(html_idx, "r") as jf:
-                data = json.load(jf)
-            for fl in data['files']:
-                found_target = fuzz_utils.approximate_python_coverage_files(
-                    function_name,
-                    data['files'][fl]['index']['relative_filename'],
-                )
-                if found_target:
-                    return cov_url + "/" + fl + ".html" + "#t" + str(lineno)
-        else:
-            logger.info("Could not find any html_status.json file")
-        return "#"
-
-    def refine_paths(self, basefolder: str) -> None:
-        """
-        Removes the project_profile's basefolder from source paths in a given profile.
-        """
-        # Only do this is basefolder is not wrong
-        if basefolder == "/":
-            return
-
-        self.fuzzer_source_file = self.fuzzer_source_file.replace(basefolder, "")
-
-        if self.function_call_depths is not None:
-            all_callsites = fuzz_cfg_load.extract_all_callsites(self.function_call_depths)
-            for cs in all_callsites:
-                cs.dst_function_source_file = cs.dst_function_source_file.replace(basefolder, "")
-
-            new_dict = {}
-            for key in self.file_targets:
-                new_dict[key.replace(basefolder, "")] = self.file_targets[key]
-            self.file_targets = new_dict
-
-    def set_all_reached_functions(self) -> None:
-        """
-        sets self.functions_reached_by_fuzzer to all functions reached
-        by LLVMFuzzerTestOneInput
-        """
-        if "LLVMFuzzerTestOneInput" in self.all_class_functions:
-            self.functions_reached_by_fuzzer = (
-                self.all_class_functions["LLVMFuzzerTestOneInput"].functions_reached
-            )
-            return
-
-        # Find Python entrypoint
-        for func_name in self.all_class_functions:
-            if "TestOneInput" in func_name:
-                reached = self.all_class_functions[func_name].functions_reached
-                self.functions_reached_by_fuzzer = reached
-                return
-
-        # TODO: make fuzz-introspector exceptions
-        raise Exception
-
-    def reaches_file(
-        self,
-        file_name: str,
-        basefolder: Optional[str] = None
-    ) -> bool:
-        if basefolder is not None:
-            new_file_name = file_name.replace(basefolder, "")
-        else:
-            new_file_name = file_name
-
-        for ff in self.file_targets:
-            logger.info(f"\t{ff}")
-
-        # Only some file paths have removed base folder. We must check for both.
-        return (file_name in self.file_targets) or (new_file_name in self.file_targets)
-
-    def reaches(self, func_name: str) -> bool:
-        return func_name in self.functions_reached_by_fuzzer
-
-    def correlate_executable_name(self, correlation_dict) -> None:
-        for elem in correlation_dict['pairings']:
-            if os.path.basename(self.introspector_data_file) in f"{elem['fuzzer_log_file']}.data":
-                self.binary_executable = str(elem['executable_path'])
-
-                lval = os.path.basename(self.introspector_data_file)
-                rval = f"{elem['fuzzer_log_file']}.data"
-                logger.info(f"Correlated {lval} with {rval}")
-
-    def get_key(self) -> str:
-        """
-        Returns the "key" we use to identify this Fuzzer profile.
-        """
-        if self.binary_executable != "":
-            return os.path.basename(self.binary_executable)
-
-        return self.fuzzer_source_file
-
-    def set_all_unreached_functions(self) -> None:
-        """
-        sets self.functions_unreached_by_fuzzer to all functiosn in self.all_class_functions
-        that are not in self.functions_reached_by_fuzzer
-        """
-        self.functions_unreached_by_fuzzer = [
-            f.function_name for f
-            in self.all_class_functions.values()
-            if f.function_name not in self.functions_reached_by_fuzzer
-        ]
-
-    def load_coverage(self, target_folder: str) -> None:
-        """Load coverage data for this profile"""
-        logger.info(f"Loading coverage of type {self.target_lang}")
-        if self.target_lang == "c-cpp":
-            self.coverage = fuzz_cov_load.llvm_cov_load(
-                target_folder,
-                self.get_target_fuzzer_filename()
-            )
-        elif self.target_lang == "python":
-            self.coverage = fuzz_cov_load.load_python_json_cov(
-                target_folder
-            )
-        else:
-            raise DataLoaderError(
-                "The profile target has no coverage loading support"
-            )
-
-    def get_target_fuzzer_filename(self) -> str:
-        return self.fuzzer_source_file.split("/")[-1].replace(".cpp", "").replace(".c", "")
-
-    def get_file_targets(self) -> None:
-        """
-        Sets self.file_targets to be a dictionarty of string to string.
-        Each key in the dictionary is a filename and the corresponding value is
-        a set of strings containing strings which are the names of the functions
-        in the given file that are reached by the fuzzer.
-        """
-        if self.function_call_depths is not None:
-            all_callsites = fuzz_cfg_load.extract_all_callsites(self.function_call_depths)
-            for cs in all_callsites:
-                if cs.dst_function_source_file.replace(" ", "") == "":
-                    continue
-                if cs.dst_function_source_file not in self.file_targets:
-                    self.file_targets[cs.dst_function_source_file] = set()
-                self.file_targets[cs.dst_function_source_file].add(cs.dst_function_name)
-
-    def get_total_basic_blocks(self) -> None:
-        """
-        sets self.total_basic_blocks to the sym of basic blocks of all the functions
-        reached by this fuzzer.
-        """
-        total_basic_blocks = 0
-        for func in self.functions_reached_by_fuzzer:
-            fd = self.all_class_functions[func]
-            total_basic_blocks += fd.bb_count
-        self.total_basic_blocks = total_basic_blocks
-
-    def get_total_cyclomatic_complexity(self) -> None:
-        """
-        sets self.total_cyclomatic_complexity to the sum of cyclomatic complexity
-        of all functions reached by this fuzzer.
-        """
-        self.total_cyclomatic_complexity = 0
-        for func in self.functions_reached_by_fuzzer:
-            fd = self.all_class_functions[func]
-            self.total_cyclomatic_complexity += fd.cyclomatic_complexity
-
-    def accummulate_profile(self, target_folder: str) -> None:
-        """
-        Triggers various analyses on the data of the fuzzer. This is used after a
-        profile has been initialised to generate more interesting data.
-        """
-        self.set_all_reached_functions()
-        self.set_all_unreached_functions()
-        self.load_coverage(target_folder)
-        self.get_file_targets()
-        self.get_total_basic_blocks()
-        self.get_total_cyclomatic_complexity()
-
-    def get_cov_uncovered_reachable_funcs(self) -> List[str]:
-        if self.coverage is None:
-            return []
-
-        uncovered_funcs = []
-        for funcname in self.functions_reached_by_fuzzer:
-            total_func_lines, hit_lines, hit_percentage = self.get_cov_metrics(funcname)
-            if total_func_lines is None:
-                uncovered_funcs.append(funcname)
-                continue
-            if hit_lines == 0:
-                uncovered_funcs.append(funcname)
-        return uncovered_funcs
-
-    def is_file_covered(
-        self,
-        file_name: str,
-        basefolder: Optional[str] = None
-    ) -> bool:
-        # We need to refine the pathname to match how coverage file paths are.
-        file_name = os.path.abspath(file_name)
-
-        # Refine filename if needed
-        if basefolder is not None and basefolder != "/":
-            new_file_name = file_name.replace(basefolder, "")
-        else:
-            new_file_name = file_name
-
-        for funcname in self.all_class_functions:
-            # Check it's a relevant filename
-            func_file_name = self.all_class_functions[funcname].function_source_file
-            if basefolder is not None and basefolder != "/":
-                new_func_file_name = func_file_name.replace(basefolder, "")
-            else:
-                new_func_file_name = func_file_name
-            if func_file_name != file_name and new_func_file_name != new_file_name:
-                continue
-            # Return true if the function is hit
-            tf, hl, hp = self.get_cov_metrics(funcname)
-            if hp is not None and hp > 0.0:
-                if func_file_name in self.file_targets or new_file_name in self.file_targets:
-                    return True
-        return False
-
-    def get_cov_metrics(
-        self,
-        funcname: str
-    ) -> Tuple[Optional[int], Optional[int], Optional[float]]:
-        if self.coverage is None:
-            return None, None, None
-        try:
-            total_func_lines, hit_lines = self.coverage.get_hit_summary(funcname)
-            if total_func_lines is None or hit_lines is None:
-                return None, None, None
-
-            hit_percentage = (hit_lines / total_func_lines) * 100.0
-            return total_func_lines, hit_lines, hit_percentage
-        except Exception:
-            return None, None, None
-
-    def write_stats_to_summary_file(self) -> None:
-        file_target_count = len(self.file_targets) if self.file_targets is not None else 0
-        fuzz_utils.write_to_summary_file(
-            self.get_key(),
-            "stats",
-            {
-                "total-basic-blocks": self.total_basic_blocks,
-                "total-cyclomatic-complexity": self.total_cyclomatic_complexity,
-                "file-target-count": file_target_count,
-            }
-        )
-
-
-class MergedProjectProfile:
-    """
-    Class for storing information about all fuzzers combined in a given project.
-
-    This means, it contains data for all fuzzers in a given project, and digests
-    the manner in a way that makes sense from a project-scope perspective. For
-    example, it does project-wide analysis of reachable/unreachable functions by
-    digesting data from all the fuzzers in the project.
-    """
-    def __init__(self, profiles: List[FuzzerProfile]):
-        self.name = None
-        self.profiles = profiles
-        self.all_functions: Dict[str, FunctionProfile] = dict()
-        self.unreached_functions = set()
-        self.functions_reached = set()
-
-        logger.info(f"Creating merged profile of {len(self.profiles)} profiles")
-        # Populate functions reached
-        logger.info("Populating functions reached")
-        for profile in profiles:
-            for func_name in profile.functions_reached_by_fuzzer:
-                self.functions_reached.add(func_name)
-
-        # Set all unreached functions
-        logger.info("Populating functions unreached")
-        for profile in profiles:
-            for func_name in profile.functions_unreached_by_fuzzer:
-                if func_name not in self.functions_reached:
-                    self.unreached_functions.add(func_name)
-
-        # Add all functions from the various profiles into the merged profile. Don't
-        # add duplicates
-        logger.info("Creating all_functions dictionary")
-        excluded_functions = {
-            "sanitizer", "llvm"
-        }
-        for profile in profiles:
-            for fd in profile.all_class_functions.values():
-                # continue if the function is to be excluded
-                if len([ef for ef in excluded_functions if ef in fd.function_name]) != 0:
-                    continue
-
-                # populate hitcount and reached_by_fuzzers and whether it has been handled already
-                for fuzzer_profile in profiles:
-                    if fuzzer_profile.reaches(fd.function_name):
-                        fd.hitcount += 1
-                        fd.reached_by_fuzzers.append(fuzzer_profile.get_key())
-                    if fd.function_name not in self.all_functions:
-                        self.all_functions[fd.function_name] = fd
-
-        # Gather complexity information about each function
-        logger.info("Gathering complexity and incoming references of each function")
-        for fp_obj in self.all_functions.values():
-            total_cyclomatic_complexity = 0
-            total_new_complexity = 0
-
-            for reached_func_name in fp_obj.functions_reached:
-                if reached_func_name not in self.all_functions:
-                    logger.error(f"Mismatched function name: {reached_func_name}")
-                    continue
-                reached_func_obj = self.all_functions[reached_func_name]
-                reached_func_obj.incoming_references.append(fp_obj.function_name)
-                total_cyclomatic_complexity += reached_func_obj.cyclomatic_complexity
-                if reached_func_obj.hitcount == 0:
-                    total_new_complexity += reached_func_obj.cyclomatic_complexity
-            if fp_obj.hitcount == 0:
-                fp_obj.new_unreached_complexity = (
-                    total_new_complexity
-                    + fp_obj.cyclomatic_complexity
-                )
-            else:
-                fp_obj.new_unreached_complexity = total_new_complexity
-            fp_obj.total_cyclomatic_complexity = (
-                total_cyclomatic_complexity
-                + fp_obj.cyclomatic_complexity
-            )
-
-        # Accumulate run-time coverage mapping
-        self.runtime_coverage = fuzz_cov_load.CoverageProfile()
-        for profile in profiles:
-            if profile.coverage is None:
-                continue
-            for func_name in profile.coverage.covmap:
-                if func_name not in self.runtime_coverage.covmap:
-                    self.runtime_coverage.covmap[func_name] = profile.coverage.covmap[func_name]
-                else:
-                    # Merge by picking highest line numbers. Here we can assume they coverage
-                    # maps have the same number of elements with the same line numbers but
-                    # different hit counts.
-                    new_line_counts = list()
-                    for idx1 in range(len(self.runtime_coverage.covmap[func_name])):
-                        try:
-                            ln1, ht1 = self.runtime_coverage.covmap[func_name][idx1]
-                            ln2, ht2 = profile.coverage.covmap[func_name][idx1]
-                        except Exception:
-                            ln1, ht1 = self.runtime_coverage.covmap[func_name][idx1]
-                            ln2, ht2 = self.runtime_coverage.covmap[func_name][idx1]
-                        # It may be that line numbers are not the same for the same function
-                        # name across different fuzzers.
-                        # This *could* actually happen, and will often (almost always) happen for
-                        # LLVMFuzzerTestOneInput. In this case we just gracefully
-                        # continue and ignore issues.
-                        if ln1 != ln2:
-                            logger.error("Line numbers are different in the same function")
-                            continue
-                        new_line_counts.append((ln1, max(ht1, ht2)))
-                    self.runtime_coverage.covmap[func_name] = new_line_counts
-            # TODO (navidem): will need to merge branch coverages (branch_cov_map) if we need to
-            # identify blockers based on all fuzz targets coverage
-        self.set_basefolder()
-        logger.info("Completed creationg of merged profile")
-
-    def get_total_complexity(self) -> Tuple[int, int]:
-        reached_complexity = 0
-        unreached_complexity = 0
-        for fd in self.all_functions.values():
-            if fd.hitcount == 0:
-                unreached_complexity += fd.cyclomatic_complexity
-            else:
-                reached_complexity += fd.cyclomatic_complexity
-        return reached_complexity, unreached_complexity
-
-    def get_total_unreached_function_count(self) -> int:
-        unreached_function_count = 0
-        for fd in self.all_functions.values():
-            if fd.hitcount == 0:
-                unreached_function_count += 1
-        return unreached_function_count
-
-    def get_total_reached_function_count(self) -> int:
-        reached_function_count = 0
-        for fd in self.all_functions.values():
-            if fd.hitcount != 0:
-                reached_function_count += 1
-        return reached_function_count
-
-    def get_all_runtime_covered_functions(self) -> List[str]:
-        all_covered_functions = []
-        for funcname in self.runtime_coverage.covmap:
-            all_covered_functions.append(funcname)
-        return all_covered_functions
-
-    def get_function_reach_percentage(self) -> float:
-        total_functions = (
-            float(self.get_total_unreached_function_count()
-                  + self.get_total_reached_function_count())
-        )
-        reached_percentage = (
-            float(self.get_total_reached_function_count() / total_functions)
-            * 100.0
-        )
-        return reached_percentage
-
-    def get_function_summaries(self) -> Tuple[int, int, int, float, float]:
-        reached_func_count = self.get_total_reached_function_count()
-        unreached_func_count = self.get_total_unreached_function_count()
-        total_functions = reached_func_count + unreached_func_count
-        reached_percentage = (float(reached_func_count) / float(total_functions)) * 100
-        unreached_percentage = (float(unreached_func_count) / float(total_functions)) * 100
-        return (
-            total_functions,
-            reached_func_count,
-            unreached_func_count,
-            reached_percentage,
-            unreached_percentage
-        )
-
-    def get_complexity_summaries(self) -> Tuple[int, int, int, float, float]:
-        complexity_reached, complexity_unreached = self.get_total_complexity()
-        total_complexity = complexity_unreached + complexity_reached
-
-        try:
-            reached_complexity_percentage = (float(complexity_reached) / (total_complexity)) * 100.0
-        except Exception:
-            logger.info("Total complexity is 0")
-            reached_complexity_percentage = 0
-        try:
-            unreached_complexity_percentage = (
-                (float(complexity_unreached) / (total_complexity)) * 100.0
-            )
-        except Exception:
-            logger.info("Total complexity is 0")
-            unreached_complexity_percentage = 0
-
-        return (
-            total_complexity,
-            complexity_reached,
-            complexity_unreached,
-            reached_complexity_percentage,
-            unreached_complexity_percentage
-        )
-
-    def write_stats_to_summary_file(self) -> None:
-        (total_complexity,
-         complexity_reached,
-         complexity_unreached,
-         reached_complexity_percentage,
-         unreached_complexity_percentage) = self.get_complexity_summaries()
-
-        fuzz_utils.write_to_summary_file(
-            "MergedProjectProfile",
-            "stats",
-            {
-                "total-complexity": total_complexity,
-                "complexity-reached": complexity_reached,
-                "complexity-unreached": complexity_unreached,
-                "reached-complexity-percentage": complexity_unreached,
-                "unreached-complexity-percentage": unreached_complexity_percentage
-            }
-        )
-
-    def set_basefolder(self) -> None:
-        """
-        Identifies a common path-prefix amongst source files in
-        This is used to remove locations within a host system to
-        essentially make paths as if they were from the root of the source code project.
-        """
-        all_strs = []
-        for f in self.all_functions.values():
-            if f.function_source_file == "/":
-                continue
-            if "/usr/include/" in f.function_source_file:
-                continue
-            all_strs.append(f.function_source_file)
-
-        self.basefolder = fuzz_utils.longest_common_prefix(all_strs)
-
-
 def read_fuzzer_data_file_to_profile(
     filename: str,
     language: str
